requirements.txt
torch~=2.5.1
flask~=3.0.3
flask-cors~=3.0.10
torchvision~=0.20.1
pillow~=10.3.0
sympy~=1.13.1
opencv-python-headless~=4.11.0.86